feat: clear file length allow list
The allow list had grown to five exceptions over time; three of those
files had already been refactored below 500 lines but were never
pruned. The remaining two — apps/be/src/games/glimworm/glimworm.service.ts
(550 lines) and apps/web/src/app/[locale]/payment/gem-success/page.tsx
(507) — needed an actual split before the list could be emptied.

glimworm.service.ts → 457
- Extracted three pure helpers into a new
  apps/be/src/games/glimworm/glimworm.service.lifecycle.ts:
  - seedWormSpawns(session, random) — the per-worm spawn-position +
    backwards-segment seeding loop pulled out of start()
  - resetWormsForLobby(session) — the per-worm clear-state loop pulled
    out of restart()
  - runGameTick(deps) — the 8-phase tick orchestrator pulled out of
    tick(), called by the class method via a deps bag (session,
    growthTargets, strategy, random, pickBotAngle, emitEvents,
    emitSnapshots). The class method now just resolves per-room state
    and forwards endSession when runGameTick reports a winner.

gem-success/page.tsx → 115
- Extracted the ANIMATIONS CSS string and every presentational sub-
  component (Shell, Icon, Title, Lede, PurpleAmount, BalanceCard,
  TokenChip, Actions, PrimaryLink, SecondaryLink, SparkleField, plus
  the private BalanceCell/Divider helpers) into a sibling
  GemSuccessComponents.tsx co-located with the route, matching the
  project's PascalCase sibling-file convention
  (PaymentSuccessView.tsx, SettingsContent.tsx, etc.).

scripts/check-file-length.js
- ALLOW_LIST set to empty. Comment refreshed to spell out the
  expectation: empty by design, exemptions need a linked refactor
  ticket.

Verification: pnpm check-file-length passes, web + BE tsc --noEmit
clean, all 85 glimworm Jest tests pass, ESLint on touched files clean.

(ARC-720)
