+++ scripts/check-file-length.js
@@ -4,15 +4,10 @@ const path = require('path');
 const MAX_LINES = 500;
 const ROOT_DIR = process.cwd();
 
-// List of files to ignore (relative to root)
-// These files currently exceed the limit and should be refactored.
-const ALLOW_LIST = new Set([
-  'apps/be/src/games/games.controller.ts',
-  'apps/be/src/games/glimworm/glimworm.service.ts',
-  'apps/web/src/widgets/CriticalGame/ui/ActiveGameView.tsx',
-  'apps/web/src/widgets/SeaBattleGame/ui/Game.tsx',
-  'apps/web/src/app/[locale]/payment/gem-success/page.tsx',
-]);
+// List of files to ignore (relative to root). Empty by design: every TS/TSX
+// file under apps/ must stay below MAX_LINES. If you genuinely need an
+// exemption, add the path here and link the follow-up refactor ticket.
+const ALLOW_LIST = new Set([]);
 
 const IGNORE_DIRS = new Set([
   'node_modules',
